requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-PyMongo==2.3.0
future==1.0.0
gunicorn==22.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pymongo==4.8.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
serial==0.0.97
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.3
